Added a reusable smoke-test toggle plus smoke-tuned benchmark defaults; the benchmark runner gains a --smoke-test flag, an env var to disable expectation evaluation, and now skips expectation enforcement in smoke (it even tries to auto-set accept_regressions via a yet-undefined helper), while harness discovery points now often return skip stubs instead of hard failures. Grace coherent memory and related early benchmarks were refactored into harness-friendly classes with workload metadata, smoke-mode iteration reductions, softer alignment checks, bandwidth metrics, and TF32/precision handling tweaks; several other models explicitly pin float32 on creation. Persistent decode lab refreshed: the “quick” knob is replaced with a smoke-test flag, decode option plumbing simplified, many scripts add harness-driven __main__ execution with adjusted timeouts/subprocess usage, and heavy CUDA/TMEM paths are intentionally disabled or converted to SKIPPED errors to avoid brittle builds. Advanced demos now guard against unsupported environments by raising SKIPPED in smoke or on missing hardware (Blackwell-only FlexAttention/FlexDecoding paths, vLLM demo scripts, speculative decoding, persistent matmul TMA, and multi-GPU/disaggregated prefills). Kernel/hardware handling tweaks: cluster pipeline examples check DSMEM capability before loading, TMA bulk copy wrappers drop strict PyTorch version gating, Float8 memory-access kernels add arch-safe loads and emit TIME_MS markers, compute-bound validation tolerances loosen under smoke, and tcgen05 extension loaders bail out in low-memory runs. New expectation snapshots were added for B200 hardware, KV-cache workloads get a lighter smoke-mode profile, and the benchmark runner can bypass expectation checks entirely via environment configuration.

+++ code/ch17/baseline_memory.py
@@ -63,7 +63,7 @@ class BaselineMemoryBenchmark(BaseBenchmark):
             nn.Linear(HIDDEN_DIM, HIDDEN_DIM),
             nn.GELU(),
             nn.Linear(HIDDEN_DIM, self.input_dim),
-        ).to(self.device).eval()
+        ).to(self.device, dtype=torch.float32).eval()
         self.host_batches = [
             torch.randint(
                 0,
@@ -87,7 +87,7 @@ class BaselineMemoryBenchmark(BaseBenchmark):
                     host_batch.add_(-0.5)
                     host_batch.mul_(2.0)
                     host_batch.tanh_()
-                    device_batch = host_batch.to(self.device, non_blocking=False)
+                    device_batch = host_batch.to(self.device, dtype=torch.float32, non_blocking=False)
                     _ = self.model(device_batch)
         self._synchronize()
